Working version with angular difference fixed in PID and boat sim. Removed simpleBoatModel and included within boatSim

+++ slap/src/iteration2/utils/nmea/nmeaDecoder.py
@@ -1,4 +1,4 @@
 class Decoder:
     def decodeAngle(self, input):
         input = input.split(",")
-        return int(input[1])
+        return float(input[1])
